moved tests from ui to commons
